New storeValues function, altered operator buttons
To be able to input one pair of numbers with an
operator and then hit new operator that will
display previous result with the new operator, I
moved the split method from the equalsButton into
new storeValues function and it is called with
every number button.
Also operator buttons now have an if statement, if
numberB is empty it will add the operator normally,
if numberB has a value it will call operator function
and add that operator value for the next calculation.