Disable and enable operator button functions added
The operator buttons and equals button are now
disabled in the html.
In the script, when first part of calculation is
entered, operator buttons are enabled with a
function, after operator is selected they are disabled
again until second part of calculation is entered.
Then operator and equals are enabled again.
Equals button is disabled again after it is clicked.