Created decimal button, added enable and disable
If numberA is empty"", a zero will be added to display
before decimal, when any operator button is clicked,
numberA is set to empty"" again, in case 2nd part
of equation starts with zero decimal.
Decimal Button is disabled after clicking it.
When an operator is clicked and disabled, decimal
button is enabled.
Equals button now calls storeValues so the result
can be checked if it has a decimal or not.
I moved the error message if statement out of the
operate function and added it to the equals button
so that when storeValues is called with equalsButton,
the operator buttons won't be enabled after
error message.