Created back button functions, fixed division char
Fixed incorrect division symbol, and removed space
around operator symbols. The splitting up of the
display is now looking for specific operator symbols
instead of empty space.
Created back button. Inside button function there
is function to pop off last character in display and
return rest of display and saved popped off character
in lastNum var for later.
If last part of display is operator, operator will
be disabled again.
If our display includes decimal and operator symbol
is deleted, decimal is disabled again.
If we delete decimal symbol, decimal is enabled
again.
If display is clear or last character in display
is operator symbol, numberA is set to "".